+++ internal/ipxe/http/middleware.go
