+++ internal/ipxe/script/auto.go
