+++ internal/ipxe/script/auto_test.go
