+++ internal/ipxe/script/custom.go
