+++ internal/ipxe/script/hook.go
