+++ internal/ipxe/script/ipxe_test.go
