+++ internal/metric/metric.go
@@ -1,4 +1,4 @@
-package metrics
+package metric
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
